client/requirements.txt
cffi==1.17.1
pycparser==2.22
PyYAML==6.0.2
RPi.GPIO==0.7.1
sounddevice==0.4.7
soundfile==0.12.1
websockets==11.0.3
